Various optimizations and refactoring

+++ mpyscenes/movie.py
@@ -23,6 +23,7 @@ class Movie:
         self.codec = "libx264"
         self.clips = []
         self.objects = []
+        self.empty_cache = True
         for k, v in kwargs.items():
             if k in self.__dict__:
                 self.__dict__[k] = v
@@ -82,7 +83,7 @@ class Movie:
             self.render_video_frames(filename=filename)
             fni = self._frame_filename(fn)
             fno = "%s.mov" % (fn)
-            cmd = [
+            cmd = (
                 "ffmpeg",
                 "-y",
                 "-r",
@@ -96,14 +97,15 @@ class Movie:
                 "-vendor",
                 "apl0",
                 "%s" % (fno),
-            ]
+            )
             r = sp.run(cmd)
-            if self.tempdir is not None:
-                tmpdir = full_path(self.tempdir + os.sep)
-                rmcmd = "rm %s*_frame*.png" % (tmpdir)
-            else:
-                rmcmd = "rm *_frame*.png"
-            r = sp.run(rmcmd, shell=True)
+            if self.empty_cache:
+                if self.tempdir is not None:
+                    tmpdir = full_path(self.tempdir + os.sep)
+                    rmcmd = "rm %s*_frame*.png" % (tmpdir)
+                else:
+                    rmcmd = "rm *_frame*.png"
+                r = sp.run(rmcmd, shell=True)
             toolboxprint(
                 "Video %d x %d @ %d fps encoded to ProRes4444 with alpha to %s"
                 % (self.width, self.height, self.fps, fno),
@@ -128,9 +130,7 @@ class Movie:
         # make a background fill for all clips if specified
         if self.bg_color is not None:
             bg = (
-                ColorClip(
-                    size=self.size, color=SceneObject.color_to_tuple(self.bg_color)
-                )
+                ColorClip(size=self.size, color=color_to_tuple(self.bg_color))
                 .set_start(tstart)
                 .set_end(tstop)
             )
@@ -140,20 +140,22 @@ class Movie:
         # generate clips as required for each frame
         for i, frame in enumerate(range(fstart, fstop)):
             t0, t1 = frame / self.fps, (frame + 1) / self.fps
-            toolboxprint(
-                "Rendering frame %3d / %3d at frame %3d | time %.3f sec"
-                % (i + 1, (fstop - fstart), frame, t0)
-            )
             clip_count = 0
+            anim_count = 0
             for obj in self.objects:
                 if not obj.is_mask:
                     clip = obj.get_clip_frame(frame, t0, t1)
                     if clip is not None:
                         self.clips.append(clip)
                         clip_count += 1
+                        anim_count += obj.animator_count
             if clip_count == 0:
                 # dummy blank frame if no active clips in frame
                 bg = ColorClip(size=self.size).set_start(t0).set_end(t1)
                 bgm = ColorClip(size=self.size, ismask=True)
                 bg = bg.set_mask(bgm).set_opacity(0)
                 self.clips.append(bg)
+            toolboxprint(
+                "Rendered frame %3d / %3d at frame %3d | time %.3f sec (%d objects, %d active animations)"
+                % (i + 1, (fstop - fstart), frame, t0, clip_count, anim_count)
+            )
